Use ImageData workaround in tests
